feat: Use cozy-scripts instead custom webpack configuration
