--- config/webpack.config.base.js
@@ -1,8 +1,0 @@
-module.exports = {
-  output: {
-    filename: '[name].bundle.js'
-  },
-  resolve: {
-    modules: ['node_modules']
-  }
-}
